+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## Version 0.25
+- MsSql cmdlets
+  - Corrected the help of Start-RscManagedVolumeSnapshot
+  - Corrected the help of Stop-RscManagedVolumeSnapshot
+  - Corrected the help of New-RscMssqlSnapshot
 
 ## Version 0.26